requirements.txt
asgiref==3.7.2
cffi==1.15.1
cryptography==41.0.3
Django==4.2.4
django-cors-headers==4.2.0
django-filter==23.2
django-rest-knox==4.2.0
djangorestframework==3.14.0
Markdown==3.4.4
psycopg2==2.9.7
pycodestyle==2.11.0
pycparser==2.21
python-decouple==3.8
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
